+++ gwbackupy/tests/mock_storage.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import gzip
 import hashlib
 import io
-from datetime import datetime
+from datetime import datetime, timezone
 from typing import IO
 
 from gwbackupy.storage.storage_interface import (
@@ -181,4 +181,4 @@ class MockStorage(StorageInterface):
 
     @staticmethod
     def __gen_mutation():
-        return str(int(datetime.utcnow().timestamp() * 1000))
+        return str(int(datetime.now(tz=timezone.utc).timestamp() * 1000))